pyproject.toml
[project]
name = "agent-two"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.6.2",
    "litellm>=1.74.4",
    "openai-agents>=0.2.2",
    "python-dotenv>=1.1.1",
]
